// resources/ts/models/UserFormInputs.ts
export interface UserFormInputs {
    firstName?: HTMLInputElement;
    lastName?: HTMLInputElement;
    email?: HTMLInputElement;
    username?: HTMLInputElement;
    password?: HTMLInputElement;
    confirmPassword?: HTMLInputElement;
    phoneNumber?: HTMLInputElement;
    role?: HTMLSelectElement;
}